Refactor trace store facade
